Added transform to move visit starts to 12:01 AM
Added a transform in `stanford.py` that will systematically adjust all visits with a start time of 12:00 AM to 12:01 AM. This means that visits should appear before any events associated with that visit when events for a given day are ordered by time. This is in harmony with the idea that clinicians and models should both have access to information about visit type when making predictions about future codes/events for a patient within a given visit.