Make `batch_size` an argument in labeler apply function (#209)
Users should be able to control the batch size. This is especially helpful when debugging and prototyping on small datasets so that you can see how long it takes to run a labeler on e.g., 10 samples at a time rather than having to wait for a whole 10k samples to process.